Refactor phone validation code into separate PhoneNumberValidations trait
